requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
chromadb>=0.3.24
sentence-transformers>=2.2.2
scikit-learn>=1.3.2
gspread>=5.7.2
oauth2client>=4.1.3
requests>=2.31.0
pyyaml
streamlit-webrtc>=0.40.0
pydub>=0.25.1
speechrecognition>=3.8.1
shopifyapi>=9.6.0
tqdm
